LayoutTests/platform/gtk/http/tests/websocket/tests/hybi/deflate-frame-parameter-expected.txt

Test WebSocket deflate-frame extension.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing query: "max_window_bits=8"
FAIL ws.extensions.search('x-webkit-deflate-frame') != -1 should be true. Was false.
FAIL ws.extensions.search('max_window_bits=8') != -1 should be true. Was false.
PASS event.data is firstMessage
PASS event.data is secondMessage
onclose() was called.
PASS closeEvent.wasClean is true
Testing query: "no_context_takeover"
FAIL ws.extensions.search('x-webkit-deflate-frame') != -1 should be true. Was false.
FAIL ws.extensions.search('no_context_takeover') != -1 should be true. Was false.
PASS event.data is firstMessage
PASS event.data is secondMessage
onclose() was called.
PASS closeEvent.wasClean is true
Testing query: "max_window_bits=8&no_context_takeover"
FAIL ws.extensions.search('x-webkit-deflate-frame') != -1 should be true. Was false.
FAIL ws.extensions.search('max_window_bits=8') != -1 should be true. Was false.
FAIL ws.extensions.search('no_context_takeover') != -1 should be true. Was false.
PASS event.data is firstMessage
PASS event.data is secondMessage
onclose() was called.
PASS closeEvent.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
